password hook abi in different file
